Move old withdrawal epoch_processing tests to block_processing
